Simplify .gitignore and refactor docker-compose configuration
This commit greatly reduces the amount of ignored directories and files in the .gitignore file, by focusing on the .idea/ directory only, to prevent polluting the repo with IDE-specific configurations. Largely, this enhances the clarity and readability of our .gitignore file. Alongside this, changes were made in the docker-compose.yaml file to ensure a better harmonization across different environments.