Update project packaging and manage snakeyaml dependency
This commit introduces several key changes in the simple-queue-sender, simple-queue-receiver, and pom.xml files. Firstly, it sets the packaging option to 'jar' in both sender and receiver pom files. This allows us to build standalone JAR files with all dependencies.

Next significant change is in the management of snakeyaml dependency. We excluded the original snakeyaml dependency from spring-boot-starter-web. This was done to manage its version independently. We then added its version 2.0 as a direct dependency.

A 'spring-cloud-aws-sqs' dependency was also removed, implying changes in how we handle SQS connections.

Lastly, in pom.xml, the 'modules' section was added to effectively manage multi-module project, including simple-queue-receiver and simple-queue-sender as modules.